fixed landing page issue
